Don't ignore tests which trigger docker exceptions
Tests were ignored if they triggered a docker container exception during start up. We should not be ignoring tests because of this and instead we should be solving the root cause of the issue. Ignoring this tests is a problem because as we've seen in the previous commit they would never run at all since the container wasn't exposing the correct ports and was failing at startup.