Add Gephi dependency
GephiTest used to not be able to run in the CI because it assumed that the real service was running locally. After much initial investigation into whether it would be possible to either reuse (or make my own) Gephi docker container, I decided not to go down the Docker route. This is because Gephi is quite an old piece of software and does not expose a programmatic way to install the GraphStreaming plugin that we need it to have.

Instead, we did the next best thing which was to create the GephiMock.